style: fix exclude again, rename types files
