+++ src/components/icon-tile-logo/types.ts
