+++ src/components/icon-tile/types.ts
@@ -1,8 +1,6 @@
 import { ProductSlug } from 'types/products'
 
-export type ProductBrandColor =
-  | 'neutral'
-  | Exclude<Exclude<ProductSlug, 'sentinel'>, 'hcp'>
+type ProductBrandColor = 'neutral' | Exclude<ProductSlug, 'sentinel' | 'hcp'>
 
 export interface IconTileProps {
   /** Pass a single child, which should be a Flight icon. For 'small' and 'medium' size, pass the 16px icon size; for other sizes pass the 24px icon size. Note that non-"color" icons will be colored using the "brandColor". */
